requirements.txt
mcp
requests
urllib3
python-dotenv
streamlit
pandas
altair
openai
starlette
uvicorn


